requirements.txt
pandas==2.0.3
SQLAlchemy==2.0.20
openpyxl==3.1.2
pyxlsb==1.0.10
psycopg2-binary==2.9.3
python-dotenv==1.0.1
matplotlib==3.9.2